Refactor com_atproto::identity
