+++ atprotolib-rs/src/types/com_atproto/identity/api_calls.rs
@@ -1,26 +1,25 @@
-use crate::{
-    api_calls::{AddApiAuth, ApiAuthConfig, ApiError},
-    types::com_atproto
-};
+use crate::api_calls::{AddApiAuth, ApiAuthConfig, ApiError};
 
-/// Describe the credentials that should be included in the DID doc of an account that is migrating to this service.
-/// 
+use super::{api_requests::{SignPlcOperationRequest, SubmitPlcOperationRequest, UpdateHandleRequest}, api_responses::{RecommendedDidCredentialsResponse, ResolveHandleResponse, SignPlcOperationResponse}};
+
+/// Describe the credentials that should be included in the DID doc of an
+/// account that is migrating to this service.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the service.
 /// * `api_auth_config` - The API authentication configuration.
 pub async fn get_recommended_did_credentials(
     host_name: &str,
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig
-) -> Result<com_atproto::identity::GetRecommendedDidCredentialsResponse, Box<dyn std::error::Error>>
+) -> Result<RecommendedDidCredentialsResponse, Box<dyn std::error::Error>>
 {
     let api_url = format!(
         "https://{}/xrpc/com.atproto.identity.getRecommendedDidCredentials",
         host_name
     );
 
-    
     let response = client
         .get(&api_url)
         .add_api_auth(api_auth_config.clone())
@@ -29,7 +28,7 @@ pub async fn get_recommended_did_credentials(
 
     match response.status() {
         reqwest::StatusCode::OK => {
-            let response_body: com_atproto::identity::GetRecommendedDidCredentialsResponse =
+            let response_body: RecommendedDidCredentialsResponse =
                 response.json().await?;
             Ok(response_body)
         }
@@ -38,9 +37,9 @@ pub async fn get_recommended_did_credentials(
 }
 
 /// Resolves a handle (domain name) to a DID.
-/// 
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the service.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `handle` - The handle to resolve.
@@ -49,7 +48,7 @@ pub async fn resolve_handle(
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig,
     handle: &str
-) -> Result<com_atproto::identity::ResolveHandleResponse, Box<dyn std::error::Error>> {
+) -> Result<ResolveHandleResponse, Box<dyn std::error::Error>> {
     let api_url = format!(
         "https://{}/xrpc/com.atproto.identity.resolveHandle",
         host_name
@@ -57,7 +56,6 @@ pub async fn resolve_handle(
 
     let query_params = vec![("handle", handle)];
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -67,7 +65,7 @@ pub async fn resolve_handle(
 
     match response.status() {
         reqwest::StatusCode::OK => {
-            let response_body: com_atproto::identity::ResolveHandleResponse =
+            let response_body: ResolveHandleResponse =
                 response.json().await?;
             Ok(response_body)
         }
@@ -75,10 +73,11 @@ pub async fn resolve_handle(
     }
 }
 
-/// Signs a PLC operation to update some value(s) in the requesting DID's document.
-/// 
+/// Signs a PLC operation to update some value(s) in the requesting DID's
+/// document.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the service.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `request` - The request to sign the PLC operation.
@@ -86,14 +85,13 @@ pub async fn sign_plc_operation(
     host_name: &str,
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig,
-    request: com_atproto::identity::SignPlcOperationRequest
-) -> Result<com_atproto::identity::SignPlcOperationResponse, Box<dyn std::error::Error>> {
+    request: SignPlcOperationRequest
+) -> Result<SignPlcOperationResponse, Box<dyn std::error::Error>> {
     let api_url = format!(
         "https://{}/xrpc/com.atproto.identity.signPlcOperation",
         host_name
     );
 
-    
     let response = client
         .post(&api_url)
         .json(&request)
@@ -103,7 +101,7 @@ pub async fn sign_plc_operation(
 
     match response.status() {
         reqwest::StatusCode::OK => {
-            let response_body: com_atproto::identity::SignPlcOperationResponse =
+            let response_body: SignPlcOperationResponse =
                 response.json().await?;
             Ok(response_body)
         }
@@ -111,10 +109,12 @@ pub async fn sign_plc_operation(
     }
 }
 
-/// Validates a PLC operation to ensure that it doesn't violate a service's constraints or get the identity into a bad state, then submits it to the PLC registry
-/// 
+/// Validates a PLC operation to ensure that it doesn't violate a service's
+/// constraints or get the identity into a bad state, then submits it to the PLC
+/// registry
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the service.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `request` - The request to submit the PLC operation.
@@ -122,14 +122,13 @@ pub async fn submit_plc_operation(
     host_name: &str,
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig,
-    request: com_atproto::identity::SubmitPlcOperationRequest
+    request: SubmitPlcOperationRequest
 ) -> Result<(), Box<dyn std::error::Error>> {
     let api_url = format!(
         "https://{}/xrpc/com.atproto.identity.submitPlcOperation",
         host_name
     );
 
-    
     let response = client
         .post(&api_url)
         .json(&request)
@@ -143,10 +142,11 @@ pub async fn submit_plc_operation(
     }
 }
 
-/// Updates the current account's handle. Verifies handle validity, and updates did:plc document if necessary. Implemented by PDS, and requires auth.
-/// 
+/// Updates the current account's handle. Verifies handle validity, and updates
+/// did:plc document if necessary. Implemented by PDS, and requires auth.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the service.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `request` - The request to update the handle.
@@ -154,14 +154,13 @@ pub async fn update_handle(
     host_name: &str,
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig,
-    request: com_atproto::identity::UpdateHandleRequest
+    request: UpdateHandleRequest
 ) -> Result<(), Box<dyn std::error::Error>> {
     let api_url = format!(
         "https://{}/xrpc/com.atproto.identity.updateHandle",
         host_name
     );
 
-    
     let response = client
         .post(&api_url)
         .json(&request)
