+++ atprotolib-rs/src/types/com_atproto/identity/api_responses.rs
@@ -1,12 +1,8 @@
 use serde::{Deserialize, Serialize};
 
-/*
-    com.atproto.identity.getRecommendedDidCredentials
-*/
-
 /// Represents a response to get recommended DID credentials.
 #[derive(Serialize, Deserialize, Debug)]
-pub struct GetRecommendedDidCredentialsResponse {
+pub struct RecommendedDidCredentialsResponse {
     /// The rotation keys.
     #[serde(rename = "rotationKeys")]
     pub rotation_keys: Vec<String>,
@@ -22,4 +18,20 @@ pub struct GetRecommendedDidCredentialsResponse {
     /// The services.
     #[serde(rename = "services")]
     pub services: serde_json::Value
+}
+
+/// Represents a response to a request to resolve a handle to a DID.
+#[derive(Serialize, Deserialize, Debug)]
+pub struct ResolveHandleResponse {
+    /// The DID associated with the handle.
+    #[serde(rename = "did")]
+    pub did: String
+}
+
+/// Represents a response to sign a PLC operation.
+#[derive(Serialize, Deserialize, Debug)]
+pub struct SignPlcOperationResponse {
+    /// A value of the operation. (?)
+    #[serde(rename = "operation")]
+    pub operation: serde_json::Value
 }
