--- atprotolib-rs/src/types/com_atproto/identity/update_handle.rs
@@ -1,13 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-/*
-    com.atproto.identity.updateHandle
-*/
-
-/// Represents a request to update a handle.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct UpdateHandleRequest {
-    /// The new handle.
-    #[serde(rename = "handle")]
-    pub handle: String
-}
